Separated matrix multiplication (DotGeneral) performance test (TestDotGeneral_PerformanceTable) behind the build tag `perf` to avoid poluting the dependencies in `go.mod`.
